requirements.txt
torch
torchvision
numpy
opencv-python
ruamel.yaml
trimesh
ultralytics
plotly
open3d-python
pyyaml
scikit-learn
tqdm
pandas
matplotlib
scipy
Pillow
